+++ src/components/Header/Header.js
@@ -69,7 +69,8 @@ class Header extends React.Component {
       showNewMessage: false,
       hideMessage: true,
       run: true,
-      arrowImg: arrowActive
+      arrowImg: arrowActive,
+      color: "#323232",
     };
   }
 
@@ -144,7 +145,7 @@ class Header extends React.Component {
     });
   }
   render() {
-    const { focus } = this.state;
+    const { focus, color } = this.state;
     const { openUsersList } = this.props;
 
     const user = JSON.parse(localStorage.getItem("user") || '{}');
@@ -168,13 +169,13 @@ class Header extends React.Component {
             />
           </NavLink>
         </NavItem>
-        <NavItem className={"d-block d-sm-none ml-auto"}>
+        <NavItem className={"d-block d-sm-none mr-auto"}>
           <img
             src={search}
             alt="search"
             width="24px"
             height="23px"
-            style={{ marginRight: 12 }}
+            style={{ marginLeft: 10 }}
           />
         </NavItem>
         <button className={`btn btn-bordered ml-auto justify-content-end d-md-down-none ${s.fullVersionBtn}`} onMouseOver={() => this.changeArrowImg()} onMouseLeave={() => this.changeArrowImgOut()}>
@@ -192,13 +193,14 @@ class Header extends React.Component {
                 width="24px"
                 height="23px"
                 className="justify-content-center"
+                style={{ opacity: '.5'}}
               />
             </InputGroupAddon>
             <Input
                 id="search-input"
                 placeholder="Search"
-                className={`${cx({ focus: !!focus})} ${s.headerSearchInput}`}
-                // style={{ borderBottomLeftRadius: 4, borderTopLeftRadius: 4}}
+                className={`${cx({ focus: !!focus, [s.headerSearchInput]: color === '#323232', [s.headerSearchInputLight]: color === '#FFFFFF'})}`}
+                //style={{ borderBottomLeftRadius: 4, borderTopLeftRadius: 4}}
             />
           </InputGroup>
         </Form>